% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculates a score from habitat suitability and distance.}
\usage{
calculate_score(hs, d)
}
\description{
Calculates a score from habitat suitability and distance.
}
