% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{get_biooracle_temperature}
\alias{get_biooracle_temperature}
\title{Reads bio-oracle temperature data.}
\usage{
get_biooracle_temperature(bbox = c(-40, 50, 20, 90))
}
\description{
Reads bio-oracle temperature data.
}
