% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\name{get_occurrence}
\alias{get_occurrence}
\title{Fetches occurrences from Eurobis for a given aphiaID}
\usage{
get_occurrence(aphiaid = 159559)
}
\description{
Fetches occurrences from Eurobis for a given aphiaID
}
