% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{get_temperature}
\alias{get_temperature}
\title{Fetches temperature data. See https://rpubs.com/markpayne/358146}
\usage{
get_temperature(bbox = c(-30, 30, 20, 90),
  time = "2015-12-15T00:00:00Z")
}
\description{
Fetches temperature data. See https://rpubs.com/markpayne/358146
}
