% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.R
\name{make_suitability}
\alias{make_suitability}
\title{Creates a habitat suitability raster based on Mahalanobis distance.}
\usage{
make_suitability(occ, temp, sal)
}
\arguments{
\item{occ}{Occurrence sf object.}

\item{temp}{Temperature raster.}

\item{sal}{Salinity raster.}
}
\description{
Creates a habitat suitability raster based on Mahalanobis distance.
}
