% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{extract_temperature}
\alias{extract_temperature}
\title{Extract temperature values for point locations.}
\usage{
extract_temperature(x, lon, lat)
}
\description{
Extract temperature values for point locations.
}
